feat: エントリーポイントを軽量化
